refactor: move enumerateTree to util
